Implement convex setting functions in convex model

+++ 01_covering_multiple_experts/experiments/src/config/config.h
@@ -14,6 +14,6 @@ public:
     double max_distance;
     uint32_t gurobi_verbosity;
     double epsilon;
-    uint32_t L;
+    double L;
     bool is_convex;
 };
